Simplify composeEnhancers declaration
Based on
https://github.com/zalmoxisus/redux-devtools-extension/tree/51c7b30d6eca4825e30491e602c925474340f9ed#12-advanced-store-setup

As shown in the README above, the redux devtools extension compose
function accepts a config object. However, our use of it did not require
any custom options and we were passing in an empty object, resulting in
unnecessarily complicated code. This commit simplifies the code by
replacing it with a one-line, non-customized version as documented in
the README.