Update scraper config to scrape hourly
There have been many complaints of outdated data. We were previously
promised in a meeting that the data will only be updated at around 4am
everyday, but this does not appear to be true as ModReg gets data faster
than that. Let's just scrape hourly to ensure that we're up to date as
well.