Add forwardRef to ExternalLink
This is a fun one. The team page was crashing, but only on the SECOND time the page is visited. Turns out React Tippy expects the wrapped component to allow refs, which doesn't work for FCs (strangely not documented anywhere it seems). Adding a forwardRef here fixed the tooltips on the page and the crash. I've also checked that every other tooltip usage seems to be valid