Use modules_v2 index
We cannot change the existing semesterData from a non-nested object
mapping to a nested one. We also don't want to simply nuke the existing
index and recreate it, as the nested object will break the existing
frontend.

We version the index instead, so that we can update the scraper without
causing errors in the scraper or the frontend. We'll remove the old
index after upgrading the frontend.

This will likely incur a small monetary cost on Elastic Cloud.