Rewrite GlobalSearchContainer tests with RTL
Prepares GlobalSearchContainer tests for functionalization. The existing
tests tested methods on the GlobalSearchContainer class, which were an
implementation detail (those methods were meant to be called by the
child `GlobalSearch` component). Those methods will be turned into
`useCallback` callbacks in a functionalized component not accessible
from outside the component, requiring a rewrite of the tests.