Don't memoize `VenueDetails` as it should be cheap to render
Cost of `memo` is probably higher than perf savings for `VenueDetails`.
All expensive variables in `VenueDetails` are memoized now, and
`Timetable` is a PureComponent, so rendering `VenueDetails` should be
cheap.

Also, `VenuesContainer` (the only component that uses `VenueDetails`)
doesn't take advantage of memoization anyway as it passes a new
`highlightPeriod` object on every render.