Force version string date to be in SGT
`new Date()` returns the current date in the build machine's local time
zone. This used to be fine for us as the production builds of NUSMods
was always produced in Singapore. However, now that NUSMods is built on
Vercel, there's no guaranteed on what the machine's time zone will be.
It is also not possible to pass in a custom `TZ` env var to the build
runner.

Solution: use the existing SGT transformer to always produce SGT dates.