Replace top/bottomBar with flex layouts
bottomBar was position:relative, which caused it to float above the page
regardless of where its parent nav bar was. This causes the tabs to
stick to the top of the screen when you rubberband-scroll on iOS.

To fix this, this commit brings the tabs inline, laying out the nav bar
using flex magic instead.

This also allows us to easily prevent the search box or other nav
bar components from overlapping with the tabs. Global search is now much
less buggy than before.