Don't memo ModulesSelect
This caused a bug where the dropdown contents won't render when the
results were updated (e.g. if you selected a mod, the dropdown won't
update). This is because the component isn't passed a prop that updates
when the results update, and it also does not have any internal state
that would be set when results update. As such, ModulesSelect cannot be
memoized.