Navbar: shift center block left
The navbar links being at the centre made it look very uneven,
especially with the search bar expanding to take up space on the right.
This commit shifts the navbar links to the left, and aligns the padding
of the navbar with the Today, Venues and Timetable components on the
medium viewport widths.